STDL-1 add the register methods using the stat management, add the navbar and the routes
