+++ src/context/AuthContext.jsx
@@ -1,0 +1,18 @@
+import React, { createContext, useState } from 'react';
+
+export const AuthContext = createContext();
+
+export const AuthProvider = ({ children }) => {
+  const [user, setUser] = useState(() => JSON.parse(localStorage.getItem('user')) || null);
+
+  const registerUser = (userData) => {
+    setUser(userData);
+    localStorage.setItem('user', JSON.stringify(userData));
+  };
+
+  return (
+    <AuthContext.Provider value={{ user, registerUser }}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
