+++ src/routes/PrivateRoute.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Navigate } from 'react-router-dom';
+import useAuth from '../hooks/useAuth';
+
+const PrivateRoute = ({ children }) => {
+  const { user } = useAuth();
+
+  return user ? children : <Navigate to="/login" />;
+};
+
+export default PrivateRoute;
